pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synoptic-charts"
version = "0.1.0"
description = "Lightweight Python package for creating synoptic meteorological charts from model data"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Keith Roberts", email = "keithrbt0@gmail.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "herbie-data>=2024.0.0",
    "matplotlib>=3.7.0",
    "cartopy>=0.22.0",
    "metpy>=1.5.0",
    "xarray>=2023.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "cfgrib>=0.9.10",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]
video = [
    "ffmpeg-python>=0.2.0",
]
cli = [
    "tqdm>=4.65.0",
]

[project.scripts]
synoptic-charts = "synoptic_charts.cli:main"

[project.urls]
Homepage = "https://github.com/krober10nd/synoptic-charts"
Repository = "https://github.com/krober10nd/synoptic-charts"
Documentation = "https://github.com/krober10nd/synoptic-charts#readme"

[tool.setuptools]
packages = ["synoptic_charts"]
